Cleanup code - moving to utils
