+++ src/screens/checkout/Checkout.css
@@ -7,8 +7,6 @@ width: 1000px !important;
 
 }
 
-
-
 .gridListTile{
 margin-top: 1%;
 width:33% !important;
