+++ src/screens/details/Details.css
@@ -28,7 +28,6 @@
 }
 
 
-
 /*CSS properties for Restaurant image and details*/
 .image{
   width: 100%;
